+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 All notable changes to `sebastianbergmann/object-reflector` are documented in this file using the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## [6.1.0] - 2026-MM-DD
+## [6.1.0] - 2026-08-13
 
 ### Changed
 
@@ -91,7 +91,7 @@ All notable changes to `sebastianbergmann/object-reflector` are documented in th
 
 * Initial release
 
-[6.1.0]: https://github.com/sebastianbergmann/object-reflector/compare/6.0.0...main
+[6.1.0]: https://github.com/sebastianbergmann/object-reflector/compare/6.0.0...6.1.0
 [6.0.0]: https://github.com/sebastianbergmann/object-reflector/compare/5.0...6.0.0
 [5.0.0]: https://github.com/sebastianbergmann/object-reflector/compare/4.0...5.0.0
 [4.0.1]: https://github.com/sebastianbergmann/object-reflector/compare/4.0.0...4.0.1